Remove need for coroutine context. Should allow java compatability
