--- lib/src/main/kotlin/konig/kontext/CoroutineKonigKontext.kt
@@ -1,8 +1,0 @@
-package org.konigsoftware.konig.kontext
-
-import com.google.protobuf.Message
-import konig.kontext.CoroutineKonigKontextKey
-import kotlin.coroutines.AbstractCoroutineContextElement
-
-internal class CoroutineKonigKontext<ContextType : Message>(val context: ContextType) :
-    AbstractCoroutineContextElement(CoroutineKonigKontextKey)
